requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.21
asyncpg==0.30.0
aiosqlite==0.20.0
python-dotenv==1.0.1
phonenumbers==8.13.52
pydantic==2.10.6
pydantic-settings==2.6.1
email-validator==2.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.22

